+++ spec/spec_helper.rb
@@ -1,14 +1,16 @@
 # This file is copied to spec/ when you run 'rails generate rspec:install'
-ENV["RAILS_ENV"] ||= 'test'
+ENV['RAILS_ENV'] ||= 'test'
 
 # Gem main file
 require 'google_analytics_mailer'
 require 'active_support/all'
 require 'email_spec'
+require 'rails/all'
+require 'rspec/rails'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
-Dir[File.join(File.dirname(__FILE__), "support/**/*.rb")].each { |f| require f }
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
 
 # Configure action mailer for test deliveries
 ActionMailer::Base.delivery_method = :test