+++ Changelog.md
@@ -1,3 +1,7 @@
+Not released changes
+
+* Extracted uri parser in its own class
+
 ## 0.1.2, released 2013-09-26
 
 * Removed blank params from generated urls